+++ test/test_interop.py
@@ -57,7 +57,10 @@ def test_dkimpy_sign(run_miltertest, private_key, dkimpy):
                 'ARC-Message-Signature',
                 IsStr(regex=r' i=2; d=example\.com; s=elpmaxe; a=rsa-sha256;\s+c=relaxed/simple; t=1234567890;\s+h=Subject:From:To;\s+(?s:.+)'),
             ],
-            ['ARC-Authentication-Results', ' i=2; example.com; arc=pass header.oldest-pass=0 smtp.remote-ip=127.0.0.1'],
+            [
+                'ARC-Authentication-Results',
+                ' i=2; example.com; \n\tarc=pass header.oldest-pass=0 smtp.remote-ip=127.0.0.1',
+            ],
         ]
     )
 
@@ -128,6 +131,9 @@ def test_perl_sign(run_miltertest, private_key, perl_mail_dkim):
                 'ARC-Message-Signature',
                 IsStr(regex=r' i=2; d=example.com; s=elpmaxe; a=rsa-sha256;\s+c=relaxed/simple; t=1234567890;\s+h=Subject:From:To;\s+(?s:.+)'),
             ],
-            ['ARC-Authentication-Results', ' i=2; example.com; arc=pass header.oldest-pass=0 smtp.remote-ip=127.0.0.1'],
+            [
+                'ARC-Authentication-Results',
+                ' i=2; example.com; \n\tarc=pass header.oldest-pass=0 smtp.remote-ip=127.0.0.1',
+            ],
         ]
     )